userspace: split policy content-rejection logic out of policies.go
Move the content-rejection concern verbatim from policies.go into a new
policies_reject.go: offendingAddressTokens, offendingApplicationTokens,
collectPolicyContentRejections, PolicyContentRejectionReasons,
policyRejectionScope, rejectionCause, addressListHasSentinel. These build
the human-readable fail-closed reasons (#3261/#3376/#4394) that mirror the
Rust integrity preflight.

The two content-rejection sentinels (unsupportedApplicationSentinel,
unsupportedAddressSentinel) stay in the residual policies.go because they
are also written by the lowering path (buildOneRuleSnapshot) — they are a
shared contract, not reject-only.

Pure code-motion, no logic change: byte-identical slices of the
origin/master policies.go. Refs #4421 (codex-173 #4).
